+++ zerver/models/users.py
@@ -814,22 +814,10 @@ class UserProfile(AbstractBaseUser, PermissionsMixin, UserBaseSettings):
         from zerver.lib.user_groups import user_has_permission_for_group_setting
         from zerver.models import Realm
 
-        if policy_name not in [
-            "can_add_custom_emoji_group",
-            "can_create_groups",
-            "can_create_private_channel_group",
-            "can_create_public_channel_group",
-            "can_create_web_public_channel_group",
-            "can_delete_any_message_group",
-            "can_delete_own_message_group",
-            "can_manage_all_groups",
-            "create_multiuse_invite_group",
-            "direct_message_initiator_group",
-            "direct_message_permission_group",
+        if policy_name not in Realm.REALM_PERMISSION_GROUP_SETTINGS and policy_name not in [
             "edit_topic_policy",
             "invite_to_stream_policy",
             "invite_to_realm_policy",
-            "move_messages_between_streams_policy",
         ]:
             raise AssertionError("Invalid policy")
 
@@ -898,7 +886,7 @@ class UserProfile(AbstractBaseUser, PermissionsMixin, UserBaseSettings):
         return self.has_permission("create_multiuse_invite_group")
 
     def can_move_messages_between_streams(self) -> bool:
-        return self.has_permission("move_messages_between_streams_policy")
+        return self.has_permission("can_move_messages_between_channels_group")
 
     def can_create_user_groups(self) -> bool:
         return self.has_permission("can_create_groups")